Added State class and cleanup for docker2
